+++ DifferentiationInterface/src/utils/basis.jl
@@ -6,7 +6,7 @@ Construct the `i`-th standard basis array in the vector space of `a`.
 function basis(a::AbstractArray{T}, i) where {T}
     b = similar(a)
     fill!(b, zero(T))
-    b[i] = one(T)
+    b[i] = oneunit(T)
     if ismutable_array(a)
         return b
     else
@@ -23,7 +23,7 @@ function multibasis(a::AbstractArray{T}, inds) where {T}
     b = similar(a)
     fill!(b, zero(T))
     for i in inds
-        b[i] = one(T)
+        b[i] = oneunit(T)
     end
     return ismutable_array(a) ? b : map(+, zero(a), b)
 end