+++ DifferentiationInterfaceTest/src/utils.jl
@@ -1,4 +1,3 @@
-mysimilar(x::Number) = one(x)
 mysimilar(x::AbstractArray) = similar(x)
 mysimilar(x::Union{Tuple,NamedTuple}) = map(mysimilar, x)